Update the project's dependencies
This change removes laminas-twilio-phone-number-validator, as I don't
know why I added it originally, and adds in the two, new required ones:
Symfony's Console Component and the WeatherAPI PHP Client.

The Console Component provides the infrastructure for creating a command
line client that can be automated with an automation tool, such as cron.
The WeatherAPI PHP Client, while limited, provides a simplistic,
OOP-based way of interacting with the WeatherAPI.