Refactor how getTrackedCities and getUsersTrackingCity return results
After some experimentation, while arrays are fine, I believe that
returning database results as hydrated entities is, from a developer
perspective, just simpler. Given that, this change refactors the above
methods to use the HydratingResultSet and the ReflectionHydrator to
hydrate a series of entity classes with the database results; making
the information easier to work with once available.